devtools/client/webide/test/test_autoselect_project.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        (async function() {
          DebuggerServer.init();
          DebuggerServer.registerAllActors();

          let win = await openWebIDE();
          let docRuntime = getRuntimeDocument(win);
          const docProject = getProjectDocument(win);

          let panelNode = docRuntime.querySelector("#runtime-panel");
          let items = panelNode.querySelectorAll(".runtime-panel-item-other");
          is(items.length, 2, "Found 2 runtime buttons");

          // Connect to local runtime
          let connectionsChanged = waitForConnectionChange("opened");
          items[1].click();

          await waitForUpdate(win, "runtime-targets");

          await connectionsChanged;
          is(Object.keys(DebuggerServer._connections).length, 1, "Locally connected");

          ok(win.AppManager.isMainProcessDebuggable(), "Main process available");

          // Select main process
          await win.Cmds.showProjectPanel();
          await waitForUpdate(win, "runtime-targets");
          SimpleTest.executeSoon(() => {
            docProject.querySelectorAll("#project-panel-runtimeapps .panel-item")[0].click();
          });

          await waitForUpdate(win, "project");

          const lastProject = Services.prefs.getCharPref("devtools.webide.lastSelectedProject");
          is(lastProject, "mainProcess:", "Last project is main process");

          connectionsChanged = waitForConnectionChange("closed");

          await nextTick();
          await closeWebIDE(win);

          await connectionsChanged;
          is(Object.keys(DebuggerServer._connections).length, 0, "Disconnected");

          // The DebuggerServer is destroyed when closing WebIDE, so re-initialize it.
          DebuggerServer.init();
          DebuggerServer.registerAllActors();

          connectionsChanged = waitForConnectionChange("opened");

          // Re-open, should reselect main process after connection
          win = await openWebIDE();

          docRuntime = getRuntimeDocument(win);

          panelNode = docRuntime.querySelector("#runtime-panel");
          items = panelNode.querySelectorAll(".runtime-panel-item-other");
          is(items.length, 2, "Found 2 runtime buttons");

          // Connect to local runtime
          items[1].click();

          await waitForUpdate(win, "runtime-targets");

          await connectionsChanged;
          is(Object.keys(DebuggerServer._connections).length, 1, "Locally connected");
          ok(win.AppManager.isMainProcessDebuggable(), "Main process available");
          is(win.AppManager.selectedProject.type, "mainProcess", "Main process reselected");

          // Wait for the toolbox to be fully loaded
          await win.UI.toolboxPromise;

          // If we happen to pass a project object targeting the same context,
          // here, the main process, the `selectedProject` attribute shouldn't be updated
          // so that no `project` event would fire.
          const oldProject = win.AppManager.selectedProject;
          win.AppManager.selectedProject = {
            type: "mainProcess",
          };
          is(win.AppManager.selectedProject, oldProject, "AppManager.selectedProject shouldn't be updated if we selected the same project");

          await win.Cmds.disconnectRuntime();

          await closeWebIDE(win);

          DebuggerServer.destroy();

          SimpleTest.finish();
        })();
      };
    </script>
  </body>
</html>
